+++ cmd/tar2ext4/tar2ext4.go
@@ -15,6 +15,7 @@ var (
 	overlay      = flag.Bool("overlay", false, "produce overlayfs-compatible layer image")
 	convertSlash = flag.Bool("convert-slash", false, "convert backslashes ('\\') in path names to slashes ('/')")
 	vhd          = flag.Bool("vhd", false, "add a VHD footer to the end of the image")
+	onlyVhd      = flag.Bool("only-vhd", false, "adds a VHD footer to the end of the file but does not convert to ext4")
 	inlineData   = flag.Bool("inline", false, "write small file data into the inode; not compatible with DAX")
 )
 
@@ -47,6 +48,9 @@ func main() {
 		}
 		if *vhd {
 			opts = append(opts, tar2ext4.AppendVhdFooter)
+		}
+		if *onlyVhd {
+			opts = append(opts, tar2ext4.OnlyAppendVhdFooter)
 		}
 		if *inlineData {
 			opts = append(opts, tar2ext4.InlineData)
